# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: SIGAT API
  version: 1.0.0
  description: Contrato inicial da API do SIGAT, alinhado aos fluxos consumidos pelo frontend.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /api/v1
    description: API versionada atrás do Nginx
...
paths:
...
  /equipment-types:
    get:
      tags: [Inventory]
      summary: Lista tipos de equipamento do catálogo padronizado
      operationId: listEquipmentTypes
      security: [{ sessionCookie: [] }]
      parameters:
        - name: includeInactive
          in: query
          required: false
          description: Incluir tipos inativos (somente DITEL).
          schema: { type: boolean }
      responses:
        '200':
          description: Tipos de equipamento dentro do escopo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EquipmentTypeListResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Inventory]
      summary: Cadastra um tipo de equipamento no catálogo padronizado
      operationId: createEquipmentType
      description: Exclusivo para administrador DITEL.
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EquipmentTypeInput'
      responses:
        '201':
          description: Tipo criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EquipmentType'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/AuditUnavailable' }
  /equipment-types/{typeId}:
    patch:
      tags: [Inventory]
      summary: Atualiza um tipo de equipamento
      operationId: updateEquipmentType
      description: Exclusivo para administrador DITEL.
      security: [{ sessionCookie: [] }]
      parameters:
        - name: typeId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EquipmentTypeInput'
      responses:
        '200':
          description: Tipo atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EquipmentType'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/AuditUnavailable' }
    delete:
      tags: [Inventory]
      summary: Desativa um tipo de equipamento
      operationId: deactivateEquipmentType
      description: Exclusivo para administrador DITEL. Desativa o tipo (não remove), preservando referências históricas de equipamentos.
      security: [{ sessionCookie: [] }]
      parameters:
        - name: typeId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Tipo desativado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EquipmentType'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/AuditUnavailable' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: sigat_session
...
  responses:
    BadRequest:
      description: Dados inválidos
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
...
    Unauthorized:
      description: Sessão ausente ou inválida
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
    Forbidden:
      description: Usuário autenticado sem uma unidade válida para acessar o dashboard da unidade
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            invalidUnitScope:
              value:
                code: FORBIDDEN
                message: Escopo de unidade obrigatório.
    NotFound:
      description: Recurso não encontrado no escopo autorizado
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
    Conflict:
      description: Conflito de unicidade ou concorrência
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
    AuditUnavailable:
      description: A mutação foi revertida porque a auditoria obrigatória não pôde ser registrada
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            auditUnavailable:
              value:
                code: AUDIT_UNAVAILABLE
                message: Não foi possível registrar auditoria.
  schemas:
...
    EquipmentType:
      type: object
      required: [id, name, description, active, createdAt, updatedAt]
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        active: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    EquipmentTypeListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/EquipmentType' }
    EquipmentTypeInput:
      type: object
      required: [name]
      properties:
        name: { type: string, maxLength: 80 }
        description: { type: string }
        active: { type: boolean }
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: EQUIPMENT_NOT_FOUND }
        message: { type: string }
        details: {}
...